requirements.txt
google-cloud-speech
google-cloud-translate
pyaudio
webrtcvad
python-dotenv
setuptools
PyQt6